share delete between TDI and DTI
